Move utils modules into `lib`
