+++ jobrunner/lib/git.py
@@ -9,8 +9,8 @@ from pathlib import Path, PurePath
 from urllib.parse import urlparse, urlunparse
 
 from jobrunner import config
-from jobrunner.string_utils import project_name_from_url
-from jobrunner.subprocess_utils import subprocess_run
+from jobrunner.lib.string_utils import project_name_from_url
+from jobrunner.lib.subprocess_utils import subprocess_run
 
 log = logging.getLogger(__name__)
