+++ jobrunner/manage_jobs.py
@@ -21,13 +21,13 @@ from jobrunner.lib import docker
 from jobrunner.lib.database import find_where
 from jobrunner.lib.git import checkout_commit
 from jobrunner.models import SavedJobRequest, State, StatusCode
-from jobrunner.path_utils import list_dir_with_ignore_patterns
+from jobrunner.lib.path_utils import list_dir_with_ignore_patterns
 from jobrunner.project import (
     get_all_output_patterns_from_project_file,
     is_generate_cohort_command,
 )
-from jobrunner.string_utils import tabulate
-from jobrunner.subprocess_utils import subprocess_run
+from jobrunner.lib.string_utils import tabulate
+from jobrunner.lib.subprocess_utils import subprocess_run
 
 log = logging.getLogger(__name__)
