+++ jobrunner/service.py
@@ -6,7 +6,7 @@ import threading
 import time
 
 from jobrunner import config, record_stats, run, sync
-from jobrunner.log_utils import configure_logging
+from jobrunner.lib.log_utils import configure_logging
 
 log = logging.getLogger(__name__)
